requirements.txt
BeautifulSoup==3.2.0
Django==1.4
EasyProcess==0.1.4
Fabric==1.4.1
PIL==1.1.7
PyVirtualDisplay==0.1.0
South==0.7.4
amqplib==0.6.1
anyjson==0.3.1
argparse==1.2.1
boto==2.3.0
celery==2.2.7
cssutils==0.9.8a3
decorator==3.3.2
distribute==0.6.19
django-bakery==0.1
django-celery==2.2.4
django-cumulus==1.0.5
django-debug-toolbar==0.9.4
django-picklefield==0.1.9
django-storages==1.1.4
django-taggit==0.9.3
entrypoint2==0.0.5
greeking==0.131
httplib2==0.7.4
kombu==1.1.6
-e git://github.com/jacobian/openstack.compute.git@89c5c6a3fcc0864e6de2a8bc10bcaf93b13a259e#egg=openstack.compute-dev
prettytable==0.5
psycopg2==2.4.2
pycrypto==2.5
pyparsing==1.5.6
python-cloudfiles==1.7.10
python-dateutil==1.5
python-memcached==1.48
pytz==2012b
selenium==2.21.1
ssh==1.7.13
webkit2png==0.8.2
wsgiref==0.1.2
